chore: rename icons path
