+++ src/core/data-display/fab-v2/fab.spec.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render, screen, waitFor } from '@testing-library/react'
 import { userEvent } from '@testing-library/user-event'
-import { IconFileCopy } from '@/icons/mui-icons'
+import { IconFileCopy } from '@/icons/mui'
 import Fab from '.'
 
 describe('Fab', () => {
