+++ src/core/feedback/validation-dialog/validation-dialog.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { CircularProgress } from '@mui/material'
 import { render, screen, fireEvent } from '@testing-library/react'
 import type { ITitles, IValidations } from '.'
-import { IconCancelOutlined, IconCheckCircleOutlined } from '@/icons/mui-icons'
+import { IconCancelOutlined, IconCheckCircleOutlined } from '@/icons/mui'
 import ValidationDialog, { ValidationStatus } from '.'
 import { theme } from '@/theme'
